Editor simulator resources and scenes, misc.: - Resources edited in Editor now opened in Simulator as well - Editor Resource and Entity objects report via signals if their data changed - Implemented resource database for editor simulator plugin - Added flags entry to type metadata - Added flag type that hides property in editor - Simulator logs now dispatched via RPC to editor - Fixed helper functions that stringify UUIDs - Added aliases for standard integer types - Refactored DataObject serialization logic, added support for serializing specific properties - Editor console entries now colored based on log type - Added UUID dialog to launcher for easier access - UUID dialog now supports converting between standard format and base64 - Some refactoring in the Variant API - Comments for future work - Misc. fixes and refactoring

+++ tools/editor/application/VadonEditor/UI/Model/Property/Numeric.cpp
@@ -41,6 +41,6 @@ namespace VadonEditor::UI
 
 	void PropertyDoubleSpinBox::value_changed(double value)
 	{
-		internal_set_value(value);
+		internal_set_value(static_cast<float>(value));
 	}
 }
